# api/doc/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Smart Lecture Hall API",
    "description": "Documentation for Smart Lecture Hall API"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Booking",
      "description": "Endpoints for bookings"
    },
    {
      "name": "User",
      "description": "Endpoints for users"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read_pets": "read your pets",
        "write_pets": "modify pets in your account"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/bookings/recent": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The most recent booking.",
            "schema": {
              "$ref": "#/definitions/Booking"
            }
          },
          "500": {
            "description": "Error retrieving booking from database."
          }
        }
      }
    },
    "/users": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Bad request."
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all users.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "array"
                },
                "items": {
                  "$ref": "#/definitions/User"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error retrieving users from database."
          }
        }
      }
    },
    "/users/{id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User ID."
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User logged in successfully."
          },
          "401": {
            "description": "Incorrect username, email or password."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/bookings": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "lh": {
                  "example": "any"
                },
                "teacherName": {
                  "example": "any"
                },
                "course": {
                  "example": "any"
                },
                "explanation": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Booking created successfully.",
            "schema": {
              "$ref": "#/definitions/Booking"
            }
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Bookings retrieved successfully.",
            "schema": {
              "$ref": "#/definitions/BookingArray"
            }
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/bookings/{id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Booking deleted successfully."
          },
          "404": {
            "description": "Booking not found."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/timetable": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Excel file saved successfully."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "get": {
        "description": "",
        "produces": [
          null
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Timetable retrieved successfully."
          },
          "404": {
            "description": "No excel file found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/feedback": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "message": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feedback saved successfully."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Feedback data retrieved successfully.",
            "schema": {
              "$ref": "#/definitions/FeedbackArray"
            }
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/feedback/{id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Feedback deleted successfully."
          },
          "404": {
            "description": "Feedback not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/bookings/all": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Bookings retrieved successfully.",
            "schema": {
              "$ref": "#/definitions/BookingArray"
            }
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/pdf/{bookingId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "PDF file retrieved successfully."
          },
          "404": {
            "description": "PDF not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/logout": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "302": {
            "description": "Redirect to the landing page."
          },
          "500": {
            "description": "Error logging out."
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "name": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            },
            "email": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            },
            "password": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        }
      }
    },
    "Booking": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "lh": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            },
            "teacherName": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            },
            "course": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            },
            "explanation": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            },
            "pdfFile": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "format": {
                          "type": "string",
                          "example": "binary"
                        }
                      }
                    },
                    "contentType": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    }
                  }
                }
              }
            },
            "createdDate": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "date-time"
                }
              }
            }
          }
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "message": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}